+++ src.xlsx/jp/co/systembase/report/renderer/xlsx/elementrenderer/LineRenderer.java
@@ -96,7 +96,7 @@ public class LineRenderer implements IElementRenderer {
 			sp.setLineStyleColor(0, 0, 0);
 			sp.setLineWidth(lineWidth);
 			if (!this.design.isNull("color")){
-				byte[] t = ColorUtil.getTriplet((String)this.design.get("color"));
+				short[] t = ColorUtil.getTriplet((String)this.design.get("color"));
 				if (t != null){
 					sp.setLineStyleColor(t[0], t[1], t[2]);
 				}
